app/requirements.txt
streamlit
psutil
pandas
plotly
st-pages
influxdb-client
py-cpuinfo
jwt